requirements.txt
fastapi
uvicorn[standard]
python-multipart
mediapipe
opencv-python
numpy
Pillow
tensorflow
pandas
pymongo
python-dotenv
pyngrok
requests
matplotlib
motor
scikit-image
scikit-learn
ffmpeg-python